- Removed Mapper and added Extensions for mapping to entities - Fixed null reference exceptions - Added unique index in junction table - Fixed response status codes

+++ UserManagement/DataAccess/Repositories/PermissionRepository.cs
@@ -69,6 +69,11 @@ namespace UserManagement.DataAccess.Repositories
         public async Task DeletePermissionAsync(int id)
         {
             var permission = await _context.Permissions.FindAsync(id);
+            if (permission == null)
+            {
+                return;
+            }
+
             _context.Permissions.Remove(permission);
             await _context.SaveChangesAsync();
         }
